+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## master
  + Add `httpClient`-Method to `LKDev\HetznerCloud\HetznerAPIClient`
+ + Remove the `backup_window`-Property from the  `enableBackups`-Method in the Class `LKDev\HetznerCloud\Models\Servers`
 
 ## 1.1.0 (14.08.2018)
  + Add `update`-Method to `\LKDev\HetznerCloud\Models\Servers\Server`-Model for easily updateing the server meta data